requirements.txt
google-api-python-client==2.131.0
google-auth==2.29.0
pdfplumber==0.11.0
firebase-admin==6.5.0
